fix: change oneclickads theme to oca theme and use default tokens
