+++ docs/storybook/themes/themes/Oca.ts
@@ -1,13 +1,9 @@
-import {
-  OneClickAdsFonts,
-  OneClickAdsIcons,
-  OneClickAdsTheme,
-} from '@ttoss/theme/OneClickAds';
+import { OcaFonts, OcaIcons, OcaTheme } from '@ttoss/theme/Oca';
 import type { ThemeObject } from '../themesObject';
 
 export const Oca: ThemeObject = {
   name: 'Oca',
-  fonts: OneClickAdsFonts,
-  icons: OneClickAdsIcons,
-  theme: OneClickAdsTheme,
+  fonts: OcaFonts,
+  icons: OcaIcons,
+  theme: OcaTheme,
 };
