+++ packages/theme/src/themes/Oca/Oca.ts
@@ -1,0 +1,5 @@
+import './OcaIcons';
+
+export { OcaTheme } from './OcaTheme';
+export { OcaFonts } from './OcaFonts';
+export { OcaIcons } from './OcaIcons';
