+++ packages/theme/src/themes/Oca/OcaFonts.ts
@@ -1,4 +1,4 @@
-export const OneClickAdsFonts = [
+export const OcaFonts = [
   'https://fonts.googleapis.com/css2?family=Outfit:wght@700&display=swap',
   'https://fonts.googleapis.com/css2?family=Source+Sans+Pro:wght@400;700&display=swap',
 ];
