+++ packages/theme/src/themes/Oca/OcaIcons.ts
@@ -1,0 +1,38 @@
+import { IconifyIcon, addIcon } from '@ttoss/react-icons';
+
+const pressureHigh: IconifyIcon = {
+  body: `
+  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
+    <path d="M16.0155 7.35948C15.8058 7.36277 15.6028 7.28706 15.4476 7.14773L11.9972 3.8762L8.54684 7.14773C8.39233 7.28912 8.1895 7.36766 7.97892 7.36766C7.76833 7.36766 7.5655 7.28912 7.411 7.14773C7.33838 7.07949 7.28055 6.99738 7.24104 6.90637C7.20153 6.81537 7.18115 6.71738 7.18115 6.61836C7.18115 6.51934 7.20153 6.42135 7.24104 6.33034C7.28055 6.23934 7.33838 6.15723 7.411 6.08898L11.4293 2.29866C11.5838 2.15728 11.7866 2.07874 11.9972 2.07874C12.2078 2.07874 12.4106 2.15728 12.5651 2.29866L16.5834 6.0784C16.6577 6.14676 16.717 6.22952 16.7575 6.3215C16.798 6.41349 16.8189 6.51274 16.8189 6.61307C16.8189 6.71339 16.798 6.81264 16.7575 6.90463C16.717 6.99661 16.6577 7.07937 16.5834 7.14773C16.4299 7.29091 16.2266 7.37048 16.0155 7.37007V7.35948Z" fill="#03FF7A"/>
+    <path d="M16.0155 12.2728C15.8058 12.2761 15.6028 12.2004 15.4476 12.0611L11.9972 8.78953L8.54684 12.0611C8.39233 12.2024 8.1895 12.281 7.97892 12.281C7.76833 12.281 7.5655 12.2024 7.411 12.0611C7.33838 11.9928 7.28055 11.9107 7.24104 11.8197C7.20153 11.7287 7.18115 11.6307 7.18115 11.5317C7.18115 11.4327 7.20153 11.3347 7.24104 11.2437C7.28055 11.1527 7.33838 11.0706 7.411 11.0023L11.4293 7.21199C11.5838 7.07061 11.7866 6.99207 11.9972 6.99207C12.2078 6.99207 12.4106 7.07061 12.5651 7.21199L16.5834 10.9917C16.6577 11.0601 16.717 11.1428 16.7575 11.2348C16.798 11.3268 16.8189 11.4261 16.8189 11.5264C16.8189 11.6267 16.798 11.726 16.7575 11.818C16.717 11.9099 16.6577 11.9927 16.5834 12.0611C16.4299 12.2042 16.2266 12.2838 16.0155 12.2834V12.2728Z" fill="#03FF7A"/>
+    <path d="M16.0155 16.9973C15.8058 17.0006 15.6028 16.9249 15.4476 16.7856L11.9972 13.514L8.54684 16.7856C8.39233 16.9269 8.1895 17.0055 7.97892 17.0055C7.76833 17.0055 7.5655 16.9269 7.411 16.7856C7.33838 16.7173 7.28055 16.6352 7.24104 16.5442C7.20153 16.4532 7.18115 16.3552 7.18115 16.2562C7.18115 16.1572 7.20153 16.0592 7.24104 15.9682C7.28055 15.8772 7.33838 15.795 7.411 15.7268L11.4293 11.9365C11.5838 11.7951 11.7866 11.7166 11.9972 11.7166C12.2078 11.7166 12.4106 11.7951 12.5651 11.9365L16.5834 15.7162C16.6577 15.7846 16.717 15.8673 16.7575 15.9593C16.798 16.0513 16.8189 16.1506 16.8189 16.2509C16.8189 16.3512 16.798 16.4505 16.7575 16.5424C16.717 16.6344 16.6577 16.7172 16.5834 16.7856C16.4299 16.9287 16.2266 17.0083 16.0155 17.0079V16.9973Z" fill="#03FF7A"/>
+    <path d="M16.0155 21.9106C15.8058 21.9139 15.6028 21.8382 15.4476 21.6989L11.9972 18.4273L8.54684 21.6989C8.39233 21.8403 8.1895 21.9188 7.97892 21.9188C7.76833 21.9188 7.5655 21.8403 7.411 21.6989C7.33838 21.6306 7.28055 21.5485 7.24104 21.4575C7.20153 21.3665 7.18115 21.2685 7.18115 21.1695C7.18115 21.0705 7.20153 20.9725 7.24104 20.8815C7.28055 20.7905 7.33838 20.7084 7.411 20.6401L11.4293 16.8498C11.5838 16.7084 11.7866 16.6299 11.9972 16.6299C12.2078 16.6299 12.4106 16.7084 12.5651 16.8498L16.5834 20.6295C16.6577 20.6979 16.717 20.7807 16.7575 20.8727C16.798 20.9646 16.8189 21.0639 16.8189 21.1642C16.8189 21.2645 16.798 21.3638 16.7575 21.4558C16.717 21.5478 16.6577 21.6305 16.5834 21.6989C16.4299 21.8421 16.2266 21.9216 16.0155 21.9212V21.9106Z" fill="#03FF7A"/>
+  </svg>
+  `,
+  width: 24,
+  height: 24,
+};
+
+export const OcaIcons: Record<
+  string,
+  | IconifyIcon
+  | {
+      __esModule: true;
+      default: IconifyIcon;
+    }
+> = {
+  optimizationOn: pressureHigh,
+};
+
+Object.entries(OcaIcons).forEach(([key, icon]) => {
+  /**
+   * "It happens when an ES module imports a CommonJS module with the default property present and the __esModule property set to true."
+   * https://github.com/evanw/esbuild/issues/1719#issuecomment-953470495
+   */
+  if ('default' in icon) {
+    addIcon(key, icon.default);
+    return;
+  }
+
+  addIcon(key, icon);
+});
