--- packages/theme/src/themes/OneClickAds/OneClickAds.ts
@@ -1,5 +1,0 @@
-import './OneClickAdsIcons.ts';
-
-export { OneClickAdsTheme } from './OneClickAdsTheme';
-export { OneClickAdsFonts } from './OneClickAdsFonts';
-export { OneClickAdsIcons } from './OneClickAdsIcons';
