--- packages/theme/src/themes/OneClickAds/tokens/borderRadius.ts
@@ -1,16 +1,0 @@
-/**
- * https://www.chakra-ui.com/docs/theming/radii
- */
-export const radii = Object.freeze({
-  none: '0',
-  '2xs': '0.0625rem',
-  xs: '0.125rem',
-  sm: '0.25rem',
-  md: '0.375rem',
-  lg: '0.5rem',
-  xl: '0.75rem',
-  '2xl': '1rem',
-  '3xl': '1.5rem',
-  '4xl': '2rem',
-  full: '9999px',
-});
