--- packages/theme/src/themes/OneClickAds/tokens/breakpoints.ts
@@ -1,10 +1,0 @@
-/**
- * https://www.chakra-ui.com/docs/theming/breakpoints
- */
-export const breakpoints = Object.freeze({
-  sm: '640px', // 40em
-  md: '768px', // 48em
-  lg: '1024px', // 64em
-  xl: '1280px', // 80em
-  '2xl': '1536px', // 96em
-});
