--- packages/theme/src/themes/OneClickAds/tokens/index.ts
@@ -1,11 +1,0 @@
-export { radii } from './borderRadius';
-export { breakpoints } from './breakpoints';
-export { sizes } from './sizes';
-export { spacing } from './spacing';
-export {
-  fontSizes,
-  letterSpacings,
-  lineHeights,
-  fontWeights,
-} from './typography';
-export { zIndices } from './zIndex';
