--- packages/theme/src/themes/OneClickAds/tokens/typography.ts
@@ -1,61 +1,0 @@
-/**
- * https://www.chakra-ui.com/docs/theming/typography#font-sizes
- */
-export const fontSizes = Object.freeze({
-  '4xs': '0.375rem', // 6px
-  '3xs': '0.5rem', // 8px
-  '2xs': '0.625rem', // 10px
-  xs: '0.75rem', // 12px
-  sm: '0.875rem', // 14px
-  md: '1rem', // 16px
-  lg: '1.125rem', // 18px
-  xl: '1.25rem', // 20px
-  '2xl': '1.5rem', // 24px
-  '3xl': '1.875rem', // 30px
-  '4xl': '2.25rem', // 36px
-  '5xl': '3rem', // 48px
-  '6xl': '3.75rem', // 60px
-  '7xl': '4.5rem', // 72px
-  '8xl': '6rem', // 96px
-  '9xl': '8rem', // 128px
-});
-
-/**
- * https://www.chakra-ui.com/docs/theming/typography#font-weights
- */
-export const fontWeights = Object.freeze({
-  hairline: 100,
-  thin: 200,
-  light: 300,
-  normal: 400,
-  medium: 500,
-  semibold: 600,
-  bold: 700,
-  extrabold: 800,
-  black: 900,
-});
-
-/**
- * https://www.chakra-ui.com/docs/theming/typography#line-heights
- */
-export const lineHeights = Object.freeze({
-  normal: 'normal',
-  none: 1,
-  shorter: 1.25,
-  short: 1.375,
-  moderate: 1.5,
-  tall: 1.625,
-  taller: '2',
-});
-
-/**
- * https://www.chakra-ui.com/docs/theming/typography#letter-spacings
- */
-export const letterSpacings = Object.freeze({
-  tighter: '-0.05em',
-  tight: '-0.025em',
-  normal: '0',
-  wide: '0.025em',
-  wider: '0.05em',
-  widest: '0.1em',
-});
