--- packages/theme/src/themes/OneClickAds/tokens/zIndex.ts
@@ -1,19 +1,0 @@
-/**
- * https://www.chakra-ui.com/docs/theming/z-index
- */
-export const zIndices = Object.freeze({
-  hide: -1,
-  auto: 'auto',
-  base: 0,
-  docked: 10,
-  dropdown: 1000,
-  sticky: 1100,
-  banner: 1200,
-  overlay: 1300,
-  modal: 1400,
-  popover: 1500,
-  skipNav: 1600,
-  toast: 1700,
-  tooltip: 1800,
-  max: 2147483647,
-});
